requirements.txt
discord==1.7.3
discord-py-slash-command==2.4.0
riotwatcher==3.1.3
python-decouple==3.4